Code refactoring: function cleaning
