+++ app_cron.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template, jsonify, request
 from psycopg2.extras import DictCursor
-from database import get_db
+from database import get_db, save_task_status
 from taskqueue import rq_queue_high
-from app_helper import save_task_status, TASK_STATUS_PENDING
+from config import TASK_STATUS_PENDING
 import uuid, time, logging
 from config import (
     TOP_N_MOODS,
