+++ app_helper_artist.py
@@ -6,7 +6,7 @@ Separated to avoid circular imports.
 
 import logging
 from database import get_db
-from tasks.memory_utils import sanitize_string_for_db
+from sanitization import sanitize_string_for_db
 
 logger = logging.getLogger(__name__)
