+++ test/unit/test_security_ssrf.py
@@ -1,4 +1,4 @@
-"""Direct unit matrix for app_helper.validate_outbound_url (SSRF guard).
+"""Direct unit matrix for ssrf_guard.validate_outbound_url (SSRF guard).
 
 The migration blueprint exercises this guard only indirectly. Here the
 allow/deny contract is pinned directly so it is a first-class, fast regression
@@ -12,7 +12,7 @@ from unittest.mock import patch
 
 import pytest
 
-from app_helper import validate_outbound_url
+from ssrf_guard import validate_outbound_url
 
 
 def _addrinfo(ip, port=80):
@@ -35,7 +35,7 @@ class TestValidateOutboundUrlSchemes:
 
     @pytest.mark.parametrize('url', ['http://8.8.8.8', 'https://8.8.8.8'])
     def test_accepts_http_and_https(self, url):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo('8.8.8.8')):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo('8.8.8.8')):
             assert validate_outbound_url(url) == (True, None)
 
 
@@ -59,7 +59,7 @@ class TestValidateOutboundUrlIpClasses:
         ('http://127.0.0.1:8096', '127.0.0.1'),
     ])
     def test_loopback_allowed(self, url, ip):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo(ip)):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo(ip)):
             assert validate_outbound_url(url) == (True, None)
 
     @pytest.mark.parametrize('url,ip', [
@@ -68,7 +68,7 @@ class TestValidateOutboundUrlIpClasses:
         ('http://192.168.1.50:8096', '192.168.1.50'),
     ])
     def test_rfc1918_allowed(self, url, ip):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo(ip)):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo(ip)):
             assert validate_outbound_url(url) == (True, None)
 
     @pytest.mark.parametrize('url,ip', [
@@ -76,7 +76,7 @@ class TestValidateOutboundUrlIpClasses:
         ('http://1.2.3.4:8096', '1.2.3.4'),
     ])
     def test_public_ip_allowed(self, url, ip):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo(ip)):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo(ip)):
             assert validate_outbound_url(url) == (True, None)
 
     @pytest.mark.parametrize('url,ip', [
@@ -87,26 +87,26 @@ class TestValidateOutboundUrlIpClasses:
         ('http://240.0.0.1', '240.0.0.1'),
     ])
     def test_dangerous_ip_classes_rejected(self, url, ip):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo(ip)):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo(ip)):
             ok, reason = validate_outbound_url(url)
         assert ok is False
         assert reason == 'Target host resolves to a disallowed IP address'
 
     def test_dns_name_resolving_to_metadata_rejected(self):
-        with patch('app_helper.socket.getaddrinfo', return_value=_addrinfo('169.254.169.254')):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=_addrinfo('169.254.169.254')):
             ok, reason = validate_outbound_url('http://metadata.internal/')
         assert ok is False
         assert reason == 'Target host resolves to a disallowed IP address'
 
     def test_unresolvable_host_rejected(self):
-        with patch('app_helper.socket.getaddrinfo', side_effect=socket.gaierror('nope')):
+        with patch('ssrf_guard.socket.getaddrinfo', side_effect=socket.gaierror('nope')):
             ok, reason = validate_outbound_url('http://does-not-exist.invalid/')
         assert ok is False
         assert reason == 'Could not resolve host'
 
     def test_resolved_invalid_ip_rejected(self):
         bad = [(socket.AF_INET, socket.SOCK_STREAM, 6, '', ('not-an-ip', 80))]
-        with patch('app_helper.socket.getaddrinfo', return_value=bad):
+        with patch('ssrf_guard.socket.getaddrinfo', return_value=bad):
             ok, reason = validate_outbound_url('http://weird.host/')
         assert ok is False
         assert reason == 'Resolved host to invalid IP'
